+++ src/extension.ts
@@ -168,6 +168,11 @@ export function activate(context: ExtensionContext) {
             SortLines.askForSortDirection();
         })
     );
+    context.subscriptions.push(
+        commands.registerTextEditorCommand("vscode-texttoolbox.InvertLines", () => {
+            SortLines.invertLines();
+        })
+    );
 
     // text manipulations
     context.subscriptions.push(